recipes - add raster-vector function, clean logic in graph funcs
Added a raster-vector function that will take in the first layer of a raster geotiff, and aggregate the values to a given list of polygons. I.e. given a population geotiff from here for example https://ghsl.jrc.ec.europa.eu/download.php?ds=pop , and a geodataframe which contains the geometries of FSAs, it will aggregate the population and add a column to the gdf with the summed pop for each polygon.
Cleaned up some logic in the graphing functions that was breaking in edge cases.
Finally, added importlib to the scaling function so users can enter whichever sklearn.preprocessing module they like, and it will still work, without having to account for all of them manually (unfortunately, the same trick doesn't work with the models, as they come from too many different packages).