recipes - auto_tune fix inlier selection, clean up logic
Code was selecting the incorrect inliers for use in the auto_tuning function, Qinglei pointed out it was selecting the points with the lowest LOF, as opposed to selecting the [cn] highest points that are not outliers (good catch! Never would have noticed that myself :) ).

Also changed the single metric LODA logic (was missing a bracket), the high_d graphing logic to condense it, and moved the contamination and neighbor variable call in the auto_tune function.